Extract extension to separate file
